Restructure local-cli/run{Windows,Wpf}
