+++ local-cli/utils/WindowsStoreAppUtils.ps1
