+++ local-cli/utils/build.js
@@ -10,10 +10,10 @@ const MSBuildTools = require('./msbuildtools');
 const Version = require('./version');
 
 function buildSolution(slnFile, buildType, buildArch) {
-  const minVersion = new Version(8, 1, 0, 0);
-  const allVersions = MSBuildTools.getAllAvailableVersions();
+  const minVersion = new Version(10, 0, 10586, 0);
+  const allVersions = MSBuildTools.getAllAvailableUAPVersions();
   if (!allVersions.some(v => v.gte(minVersion))) {
-    throw new Error('Must have a minimum Windows SDK version 8.1 installed');
+    throw new Error('Must have a minimum Windows SDK version 10.0.10586.0 installed');
   }
 
   console.log(chalk.green(`Building ${slnFile}`));
@@ -23,13 +23,13 @@ function buildSolution(slnFile, buildType, buildArch) {
 
 function restoreNuGetPackages(options, slnFile) {
   console.log(chalk.green('Restoring NuGet packages'));
-  const nugetPath = options.nugetPath || path.join(options.root, 'node_modules/react-native-windows/local-cli/runWpf/.nuget/nuget.exe');
+  const nugetPath = options.nugetPath || path.join(options.root, 'node_modules/react-native-windows/local-cli/.nuget/nuget.exe');
   const results = execSync(`"${nugetPath}" restore "${slnFile}" -NonInteractive`).toString().split(EOL);
   results.forEach(result => console.log(chalk.white(result)));
 }
 
 function getSolutionFile(options) {
-  return glob.sync(path.join(options.root, 'wpf/*.sln'))[0];
+  return glob.sync(path.join(options.root, 'windows/*.sln'))[0];
 }
 
 module.exports = {
