+++ local-cli/utils/msbuildtools.js
